% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_merge_terminal.R
\name{merge_terminal_sales}
\alias{merge_terminal_sales}
\title{Merge terminal and sales/purchase data}
\usage{
merge_terminal_sales(sales_path, terminal_path)
}
\arguments{
\item{sales_path}{path of sales/purchasing data (cleaned)}

\item{terminal_path}{path of terminal data (cleaned)}
}
\description{
This function is the last step in the cleaning process
}
\examples{
merge_terminal_sales("C:/Users/Documents/Data/cleaned_sales_data.xlsx", "C:/Users/Documents/Data/cleaned_terminal_data.xlsx")
}
\keyword{cats}
