% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_clean.R
\name{process_purchase_sales_data}
\alias{process_purchase_sales_data}
\title{Process purchasing and sales data}
\usage{
process_purchase_sales_data(pur, sal)
}
\arguments{
\item{pur}{the path of the purchase history data (one dataset only, xlsx format) as a string}

\item{sal}{the path of the sales history data (one dataset only, xlsx format) as a string}
}
\description{
This function is the first step in the data cleaning process
}
\examples{
process_purchase_sales_data(pur = "C:/Users/Documents/Data/Purchase History/Pur_History_2014-2018.xlsx",
                            sal = "C:/Users/Documents/Data/Sales History/Sales Data 2014-2018.xlsx")

}
\keyword{history}
\keyword{purchase}
\keyword{sales}
