% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_terminal_processing.R
\name{process_terminal}
\alias{process_terminal}
\title{Process terminal data}
\usage{
process_terminal(path)
}
\arguments{
\item{path}{of the data, as a string. Needs to point to a xlsx file (only one)}
}
\description{
This function allows you to process terminal data
}
\keyword{processing}
\keyword{terminal}
